+++ upsetplot/reformat.py
@@ -150,6 +150,8 @@ class QueryResult:
     data : DataFrame
         Selected samples. The index is a MultiIndex with one boolean level for
         each category.
+    subsets : dict[frozenset, DataFrame]
+        Dataframes for each intersection of categories.
     subset_sizes : Series
         Total size of each selected subset as a series. The index is as
         for `data`.
@@ -165,12 +167,22 @@ class QueryResult:
         return ("QueryResult(data={data}, subset_sizes={subset_sizes}, "
                 "category_totals={category_totals}".format(**vars(self)))
 
+    @property
+    def subsets(self):
+        categories = np.asarray(self.data.index.names)
+        return {
+            frozenset(categories.take(mask)): subset_data
+            for mask, subset_data
+            in self.data.groupby(level=list(range(len(categories))),
+                                 sort=False)
+        }
+
 
 def query(data, present=None, absent=None,
           min_subset_size=None, max_subset_size=None,
           min_degree=None, max_degree=None,
           sort_by='degree', sort_categories_by='cardinality',
-          subset_size='auto', sum_over=None):
+          subset_size='auto', sum_over=None, include_empty_subsets=False):
     """Transform and filter a categorised dataset
 
     Retrieve the set of items and totals corresponding to subsets of interest.
@@ -199,16 +211,19 @@ def query(data, present=None, absent=None,
         Minimum degree of a subset to be reported.
     max_degree : int, optional
         Maximum degree of a subset to be reported.
-    sort_by : {'cardinality', 'degree', None}
+    sort_by : {'cardinality', 'degree', '-cardinality', '-degree',
+               'input', '-input'}
         If 'cardinality', subset are listed from largest to smallest.
         If 'degree', they are listed in order of the number of categories
-        intersected. If None, the order they appear in the data input is
+        intersected. If 'input', the order they appear in the data input is
         used.
+        Prefix with '-' to reverse the ordering.
 
         Note this affects ``subset_sizes`` but not ``data``.
-    sort_categories_by : {'cardinality', None}
+    sort_categories_by : {'cardinality', '-cardinality', 'input', '-input'}
         Whether to sort the categories by total cardinality, or leave them
-        in the provided order.
+        in the input data's provided order (order of index levels).
+        Prefix with '-' to reverse the ordering.
     subset_size : {'auto', 'count', 'sum'}
         Configures how to calculate the size of a subset. Choices are:
 
@@ -227,6 +242,9 @@ def query(data, present=None, absent=None,
         If `subset_size='sum'` or `'auto'`, then the intersection size is the
         sum of the specified field in the `data` DataFrame. If a Series, only
         None is supported and its value is summed.
+    include_empty_subsets : bool (default=False)
+        If True, all possible category combinations will be returned in
+        subset_sizes, even when some are not present in data.
 
     Returns
     -------
@@ -268,6 +286,21 @@ def query(data, present=None, absent=None,
            True   False     3
     False  True   False     1
     Name: size, dtype: int64
+    >>>
+    >>> # Getting each subset's data
+    >>> result = query(data)
+    >>> result.subsets[frozenset({"cat1", "cat2"})]
+                index     value
+    cat1  cat2 cat0
+    False True False      3  1.333795
+    >>> result.subsets[frozenset({"cat1"})]
+                        index     value
+    cat1  cat2  cat0
+    False False False      5  0.918174
+                False      8  1.948521
+                False      9  1.086599
+                False     13  1.105696
+                False     19  1.339895
     """
 
     data, agg = _aggregate_data(data, subset_size, sum_over)
@@ -275,6 +308,20 @@ def query(data, present=None, absent=None,
     totals = [agg[agg.index.get_level_values(name).values.astype(bool)].sum()
               for name in agg.index.names]
     totals = pd.Series(totals, index=agg.index.names)
+
+    if include_empty_subsets:
+        nlevels = len(agg.index.levels)
+        if nlevels > 10:
+            raise ValueError(
+                "include_empty_subsets is supported for at most 10 categories")
+        new_agg = pd.Series(0,
+                            index=pd.MultiIndex.from_product(
+                                [[False, True]] * nlevels,
+                                names=agg.index.names),
+                            dtype=agg.dtype,
+                            name=agg.name)
+        new_agg.update(agg)
+        agg = new_agg
 
     data, agg = _filter_subsets(data, agg,
                                 min_subset_size=min_subset_size,
@@ -284,21 +331,31 @@ def query(data, present=None, absent=None,
                                 present=present, absent=absent)
 
     # sort:
-    if sort_categories_by == 'cardinality':
-        totals.sort_values(ascending=False, inplace=True)
-    elif sort_categories_by is not None:
+    if sort_categories_by in ('cardinality', '-cardinality'):
+        totals.sort_values(ascending=sort_categories_by[:1] == '-',
+                           inplace=True)
+    elif sort_categories_by == '-input':
+        totals = totals[::-1]
+    elif sort_categories_by in (None, 'input'):
+        pass
+    else:
         raise ValueError('Unknown sort_categories_by: %r' % sort_categories_by)
     data = data.reorder_levels(totals.index.values)
     agg = agg.reorder_levels(totals.index.values)
 
-    if sort_by == 'cardinality':
-        agg = agg.sort_values(ascending=False)
-    elif sort_by == 'degree':
+    if sort_by in ('cardinality', '-cardinality'):
+        agg = agg.sort_values(ascending=sort_by[:1] == '-')
+    elif sort_by in ('degree', '-degree'):
         index_tuples = sorted(agg.index,
-                              key=lambda x: (sum(x),) + tuple(reversed(x)))
+                              key=lambda x: (sum(x),) + tuple(reversed(x)),
+                              reverse=sort_by[:1] == '-')
         agg = agg.reindex(pd.MultiIndex.from_tuples(index_tuples,
                                                     names=agg.index.names))
-    elif sort_by is None:
+    elif sort_by == '-input':
+        print("<", agg)
+        agg = agg[::-1]
+        print(">", agg)
+    elif sort_by in (None, 'input'):
         pass
     else:
         raise ValueError('Unknown sort_by: %r' % sort_by)